Clean up header organization and fix build errors
Major changes:
- Move extern declarations from .c files to proper headers per CLAUDE.md guidelines
- Add missing includes (mmu/mmu.h, time/time.h, misc/misc.h, cal/cal.h) across subsystems
- Fix function signature mismatches between headers and implementations:
  - MST_$MAP_CANNED_AT: 8 parameters
  - MST_$DISKLESS_INIT: 3 parameters
  - MST_$WIRE_AREA: 5 parameters
  - VOLX_$MOUNT: 9 parameters
  - DBUF_$GET_BLOCK: 6 parameters
  - FILE_$LOCK: 6 parameters
  - PROC2_$INIT: int32_t first parameter
  - FUN_00e29138: int8_t return type with void* params
- Remove duplicate/conflicting declarations from os_internal.h that conflict with included headers
- Fix CRASH_SYSTEM calls to pass pointers (&status) instead of values
- Fix TIME_$WAIT calls to use correct 2-parameter signature
- Fix type conflicts between headers (NETWORK_$DISKLESS int8_t, PROC1_$CURRENT uint16_t)
- Remove NETLOG_$LOG_IT duplicate from pmap.h (declared in ast.h)
- Add missing MMAP internal globals to mmap_internal.h
- Add SMD and TPAD subsystem files

Note: pmap/init_ws_scan.c still has decompilation issues requiring further analysis.

+++ domain_os/cal/verify.c
@@ -61,7 +61,7 @@ char CAL_$VERIFY(int *max_allowed_delta, void *param_2, char *param_3, status_$t
         do {
             FUN_00e825f4("Do you want to run DOMAIN/OS with the current calendar setting? (Y/N) ",
                          param_2);
-            TERM_$READ((void *)0xe684a4, input, (void *)0xe684a6, &status);
+            TERM_$READ((void *)0xe684a4, (unsigned int)(uintptr_t)input, (void *)0xe684a6, &status);
 
             if (input[0] == 'Y' || input[0] == 'y') {
                 return 0xFF;
